Fix the H2 server lifecycle when sql expressions are provided in the connection specification
