+++ legend-engine-executionPlan-execution-store-relational-connection/src/test/java/org/finos/legend/engine/plan/execution/stores/relational/connection/test/TestConnectionObjectProtocol_server.java
@@ -1,0 +1,53 @@
+// Copyright 2021 Goldman Sachs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package org.finos.legend.engine.plan.execution.stores.relational.connection.test;
+
+import org.finos.legend.engine.plan.execution.stores.relational.connection.RelationalExecutorInfo;
+import org.finos.legend.engine.plan.execution.stores.relational.connection.authentication.strategy.SnowflakePublicAuthenticationStrategy;
+import org.finos.legend.engine.plan.execution.stores.relational.connection.driver.vendors.snowflake.SnowflakeManager;
+import org.finos.legend.engine.plan.execution.stores.relational.connection.ds.specifications.SnowflakeDataSourceSpecification;
+import org.finos.legend.engine.plan.execution.stores.relational.connection.ds.specifications.keys.SnowflakeDataSourceSpecificationKey;
+import org.finos.legend.engine.shared.core.vault.PropertiesVaultImplementation;
+import org.finos.legend.engine.shared.core.vault.Vault;
+import org.junit.Test;
+
+import javax.security.auth.Subject;
+import java.io.FileInputStream;
+import java.util.Properties;
+
+public class TestConnectionObjectProtocol_server extends org.finos.legend.engine.plan.execution.stores.relational.connection.test.DbSpecificTests
+{
+    @Override
+    protected Subject getSubject()
+    {
+        return null;//SubjectTools.getLocalSubject();
+    }
+
+    @Test
+    public void testSnowflakePublicConnection() throws Exception
+    {
+        Properties properties = new Properties();
+        properties.load(new FileInputStream("../legend-engine-server/src/test/resources/org/finos/legend/engine/server/test/snowflake.properties"));
+        Vault.INSTANCE.registerImplementation(new PropertiesVaultImplementation(properties));
+
+        SnowflakeDataSourceSpecification ds =
+                new SnowflakeDataSourceSpecification(
+                        new SnowflakeDataSourceSpecificationKey("ki79827", "us-east-2", "LEGENDRO_WH", "KNOEMA_RENEWABLES_DATA_ATLAS", "aws"),
+                        new SnowflakeManager(),
+                        new SnowflakePublicAuthenticationStrategy("SF_KEY", "SF_PASS", "LEGEND_RO_PIERRE"),
+                        new RelationalExecutorInfo());
+        testConnection(ds::getConnectionUsingSubject, "select * from KNOEMA_RENEWABLES_DATA_ATLAS.RENEWABLES.DATASETS");
+    }
+}
