+++ legend-engine-executionPlan-execution-store-relational/src/test/java/org/finos/legend/engine/plan/execution/stores/relational/test/stream/TestStreamInput.java
@@ -15,7 +15,7 @@
 package org.finos.legend.engine.plan.execution.stores.relational.test.stream;
 
 import org.finos.legend.engine.plan.execution.stores.relational.AlloyH2Server;
-import org.finos.legend.engine.plan.execution.stores.relational.utils.DynamicPortGenerator;
+import org.finos.legend.engine.shared.core.port.DynamicPortGenerator;
 import org.finos.legend.engine.plan.execution.stores.relational.plugin.Relational;
 import org.finos.legend.engine.plan.execution.stores.relational.plugin.RelationalStoreState;
 import org.finos.legend.engine.plan.execution.stores.relational.result.RelationalResult;
